+++ src/main/java/de/srendi/advancedperipherals/common/data/DataGenerators.java
@@ -25,6 +25,8 @@ public class DataGenerators {
         generator.addProvider(event.includeServer(), new TurtleUpgradesProvider(generator));
         generator.addProvider(event.includeServer(), new PocketUpgradesProvider(generator));
         generator.addProvider(event.includeServer(), new PoiTypeProvider(generator, existingFileHelper));
+
+        generator.addProvider(event.includeClient(), new EnUsLanguageProvider(generator));
     }
 
 }